Improve assignments

+++ property-based-testing/unique/solution/MathArraysPBTest.java
@@ -1,28 +1,27 @@
 package delft;
 
-import net.jqwik.api.ForAll;
-import net.jqwik.api.Property;
-import net.jqwik.api.constraints.DoubleRange;
-import net.jqwik.api.constraints.IntRange;
-import net.jqwik.api.constraints.Size;
-import net.jqwik.api.constraints.UniqueElements;
-
-import java.util.Comparator;
-import java.util.List;
-import java.util.Random;
-
 import static org.assertj.core.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.*;
+
+import java.util.*;
+import java.util.stream.*;
+import net.jqwik.api.*;
+import net.jqwik.api.arbitraries.*;
+import net.jqwik.api.constraints.*;
+import org.junit.jupiter.api.*;
+import org.junit.jupiter.params.*;
+import org.junit.jupiter.params.provider.*;
 
 class MathArraysPBTest {
 
     @Property
-    // an array of size 100 with doubles between [1,20] will definitely have repeated numbers
+    // an array of size 100 with integers between [1,20] will definitely have repeated numbers
     // if you use only @ForAll, you are likely to never have repeated elements in the array, and thus,
     // never exercising the full property!
     void unique( 
             @ForAll
             @Size(value = 100)
-            List<@DoubleRange(min = 1, max = 20) Double> numbers) {
+            List<@IntRange(min = 1, max = 20) Integer> numbers) {
 
         double[] doubles = convertListToArray(numbers);
         double[] result = MathArrays.unique(doubles);
@@ -34,7 +33,7 @@ class MathArraysPBTest {
     }
 
     /** Use this method to convert a list of integers to an array */
-    private double[] convertListToArray(List<Double> numbers) {
+    private double[] convertListToArray(List<Integer> numbers) {
         double[] array = numbers.stream().mapToDouble(x -> x).toArray();
         return array;
     }
